Actualización de archivos del proyecto

+++ src/main/java/cl/duoc/ferremasapp/controller/BancoCentralController.java
@@ -3,6 +3,7 @@ package cl.duoc.ferremasapp.controller;
 import cl.duoc.ferremasapp.service.BancoCentralService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import java.math.BigDecimal;
 
 @RestController
 @RequestMapping("/api/banco-central")
@@ -13,11 +14,13 @@ public class BancoCentralController {
 
     @GetMapping("/tasa-cambio")
     public String obtenerTasaCambio() {
-        return bancoCentralService.obtenerTasaCambio();
+        return bancoCentralService.obtenerTasaCambio(); // Retorna el String directo
     }
 
     @GetMapping("/convertir")
-    public double convertirMonto(@RequestParam double montoExtranjero, @RequestParam double tasaCambio) {
+    public BigDecimal convertirMonto(
+            @RequestParam BigDecimal montoExtranjero,
+            @RequestParam BigDecimal tasaCambio) {
         return bancoCentralService.convertirMonto(montoExtranjero, tasaCambio);
     }
 }
